Refactor to support QR Codes as standalone app
