+++ config/terraform/aws/cloudfront.tf
@@ -54,7 +54,7 @@ resource "aws_cloudfront_distribution" "maintenance_mode" {
     }
   }
   viewer_certificate {
-    acm_certificate_arn      = aws_acm_certificate.covidportal_maintenance.arn
+    acm_certificate_arn      = aws_acm_certificate_validation.cert.certificate_arn
     minimum_protocol_version = "TLSv1.2_2019"
     ssl_support_method       = "sni-only"
   }
