+++ portal/templates/403.html
@@ -1,5 +1,5 @@
 {% extends 'base.html' %}
-{% load i18n %}
+{% load i18n settings %}
 
 {% block title %}
     {% trans "No entry" %}
@@ -10,10 +10,12 @@
     <h1>{% trans "No entry" %}</h1>
     <p>{% trans "You do not have permission to visit this page." %}</p>
     <div class="content-button push-down">
-        {% if user.is_authenticated %}
-            <a href="{% url 'start' %}" role='button' draggable='false'>{% trans "New patient" %}</a>
-        {% else %}
-            <a href="{% url 'login' %}" role='button' draggable='false'>{% trans "Log in" %}</a>
+        {% if 'APP_SWITCH'|get_setting != 'QRCODE' %}
+            {% if user.is_authenticated %}
+                <a href="{% url 'start' %}" role='button' draggable='false'>{% trans "New patient" %}</a>
+            {% else %}
+                <a href="{% url 'login' %}" role='button' draggable='false'>{% trans "Log in" %}</a>
+            {% endif %}
         {% endif %}
     </div>
 {% endblock %}
