+++ portal/templates/includes/footer.html
@@ -1,13 +1,15 @@
-{% load static %}
+{% load static settings %}
 {% load i18n %}
 
 <footer>
     <div class="page--container">
         <div class="footer-links">
             <ul>
-                <li><a href="{% url 'contact:index' %}">{% trans "Contact us" %}</a></li>
-                <li><a href="{% url 'privacy' %}">{% trans "Privacy" %}</a></li>
-                <li><a href="{% url 'terms' %}">{% trans "Terms of use" %}</a></li>
+                {% if 'APP_SWITCH'|get_setting != 'QRCODE' %}
+                    <li><a href="{% url 'contact:index' %}">{% trans "Contact us" %}</a></li>
+                    <li><a href="{% url 'privacy' %}">{% trans "Privacy" %}</a></li>
+                    <li><a href="{% url 'terms' %}">{% trans "Terms of use" %}</a></li>
+                {% endif %}
             </ul>
         </div>
         <div class="canada-wordmark"><img src="{% static 'img/wmms-blk.svg' %}" alt="{% trans 'Symbol of the Government of Canada' %}" /></div>
