+++ portal/templates/includes/nav_header.html
@@ -1,17 +1,19 @@
-{% load i18n %}
+{% load i18n settings %}
 {% load language_switcher %}
 <nav class="nav--header" aria-label="{% trans 'Language and account navigation' %}">
   <ul>
-    {% if user.is_authenticated %}
-      {# "user.is_authenticated" means they have passed the initial login screen (email, password) #}
-      {# "user.is_verifed" means they have also entered a 2FA code #}
-      {% if user.is_verified %}
-        <li><a href="{% url 'user_profile' user.id %}">{{ _("Manage your account") }}</a></li>
+      {% if 'APP_SWITCH'|get_setting != 'QRCODE' %}
+        {% if user.is_authenticated %}
+          {# "user.is_authenticated" means they have passed the initial login screen (email, password) #}
+          {# "user.is_verifed" means they have also entered a 2FA code #}
+          {% if user.is_verified %}
+            <li><a href="{% url 'user_profile' user.id %}">{{ _("Manage your account") }}</a></li>
+          {% endif %}
+          <li><a href="{% url 'logout' %}">{% trans 'Log out' %}</a></li>
+        {% elif request.resolver_match.url_name != 'login' %}
+          <li><a href="{% url 'login' %}">{% trans 'Log in' %}</a></li>
+        {% endif %}
       {% endif %}
-      <li><a href="{% url 'logout' %}">{% trans 'Log out' %}</a></li>
-    {% elif request.resolver_match.url_name != 'login' %}
-      <li><a href="{% url 'login' %}">{% trans 'Log in' %}</a></li>
-    {% endif %}
     <li>
         {% get_current_language as LANGUAGE_CODE %}
         {% get_language_info for "en" as langEn %}
