+++ register/tests.py
@@ -9,7 +9,7 @@ from .models import Registrant
 
 class RegisterView(TestCase):
     def setUp(self):
-        Switch.objects.create(name="QR_CODES", active=True)
+        pass
 
     def test_start_page(self):
         response = self.client.get(reverse("register:start"))
